----------------------------------------------------------------------  Cyclic Boundary condition Implementation
a) Performed send_recv's of the ICBC_FLAG and FLAG variables. This way the ghost layers have the exact flags as the cells which they represent.

b) The calc_resid and dot_product(in leq_bicgs.f) are modified to refer to the internal cells of the domain excluding the boundaries.

c) source_*_g(s).f are modfied so that the pressure drop conditions are correctly prescribed. In the cyclic decomposed case, all the cells are fluid across the cyclic boundary point and this has been overcomed by using actual location instead of depending on cyclic_at_e condition.

d) A bug write_ab_m.f is fixed and now the routine is more efficient. Earlier, there were communications being performed under the do_loop and that was killing the routine - now those communications are moved outside the loop.

e) A new routine write_ab_m_var.f is created and added to the repository. This is to pass another variable in addition to a & b and this helps in debugging.

f) A modification has been made to get_flow_bc.f to extend the boundaries into the ghost layers for the cylic decomposed case.

g) gridmap_mod.f has been modfied to allow the cyclic decomposition and to not allow decomposition in j-direction.

h) function.inc defines the new variable "is_on_myPE_wobnd". This refers to all the cells on the PE without boundaries.

i) The averaging routines have been modified so that only the interior cells are used.

j) Two layers are communicated in the msolve routines to yield better preconditioner. If this is going to be costly for our applications we can revert back to communicating one layer and in that case it is going to take more number of iterations.

 Committing in .

 Modified Files:
 	calc_resid.f check_data_06.f check_data_07.f corner.f
 	function.inc get_flow_bc.f get_walls_bc.f leq_bicgs.f
 	machine.f mfix_l.make mfix_u.make set_flags.f set_index1a.f
 	source_u_g.f source_u_s.f source_v_g.f source_v_s.f
 	source_w_g.f source_w_s.f vavg_u_g.f vavg_u_s.f vavg_v_g.f
 	vavg_v_s.f vavg_w_g.f vavg_w_s.f write_ab_m.f
 	dmp_modules/gridmap_mod.f
 Added Files:
 	OSF1.F
 ----------------------------------------------------------------------