Changes associated with stiff chemistry solver:
1. Removed duplicate variables associated with variable density model.
2. Removed reaction rate calculations prior to time loop.
3. Removed coeff module as most variables were redefined locally.
4. Removed copy/replace of SUM_R_g/s variables (unnecessary).
5. Modified call to stiff chemistry solver (time_march).
6. Modified reaction restart for stiff solver.

To Do:
Additional work on the stiff solver is needed. This check-in contains
the majority of the changes to files outside of the stiff solver routines.

The reaction restart will fail. A better approach is needed so that when
running in DMP, that all processes adjust the time step and respond
accordingly.

The stiff solver implementation does not contain the variable density
model. This needs to be included.