modified bc_theta routine for better clarity
Modification in bc_theta causes failure in the ahmadi model test case.
The problem was traced to a slightly different orderng in calculation of
the variables hw and cw.  The subsequent difference in these variables
appears in the 15-16th decimal place on some iterations in some time
steps (occasionally appearing in the 14th decimal place). This small
difference is enough to cause failure in the test case in the 1667th time
step (with DT=0.1). If the time step is reduced by half (DT=0.05) the
case will run to completion. To facilitate future development the old
method is being replaced and the mfix.dat updated with a slightly
smaller time step.