Changes to DES convection implementation
Particle-gas convective heat transfer was changed to reflect the
way species and heat of reaction between the phases are mananged.
Namely, the heat exchanged to the particles is integrated over the
DEM time steps and applied as an explict source to the continuum
(gas) phase.

Note, it is possible (and would be fairly easy) to directly modify
the gas phase temperature at the end of each DEM time step. This
would loosely follow how the stiff solver works.