Added a buffer to VSD cut-off switch
Using DIL_EP_s as the limiter switch for the VSD model can cause
problems mid-iteration wereby solids go from below to above the
value. Thus, a padding was added to prevent problems.  The pad
could be made tigher if needed. However, it shouldn't have any
noticable effect on simulations as this is only an issue for
very dilute solids.