Fix Issue #23: Failed runs with Intel
There was a format statement issue that would cause some Intel runs
to fail with executables built with debug flags. The change keeps
the current string format but pulls the file extention from a sting.

This is consistent to how the SPx files are opened and should not
cause any read issues.