Fixed some OMP directives with some cleanup
Some of the OMP directives were broken causing the FluidBed_DES
case to not fluidize. This was the result of some missing
reductions when calculating DES_ROP_s (etc).

Some of the other aspects of this routine were also cleaned when
tracking down the issue. There is still a lot of changes that
could be implemented to simplify the code.