Changes in explicit drag calculation for DMP mode
o The pressure graident shouldn't be calculated and applied to the
  particle until the end of the fuild time step so that the
  converged pressure values are used.

o Interpolation weights were originally calculated only for cells
  owned by the current rank. That is not necessary and causes
  errors in DMP mode as values interpolated to a particle's postion
  will not take into account data from adjacent cells owned by
  another rank.

o The call to usr0_des was moved before the other function calls
  in des_time_march.

o Ghost particle volumes are now calculated. This infomration
  is needed to ensure an accurate solids volume fraction. This
  data could be sent/recv via MPI but it is easily calculated
  and therefore doesn't need to increase the buffer size.

o Ghost particles are now binned to the fluid grid too. This
  is needed to ensure an accurate solids volume fraction calculation.
  The original approach used it's own MPI routines for this, however
  the new approach should be faster and isn't as complex.
  Note: ghost particles are still omitted from this list of particles
  in the fluid cell (PIJK and PIC).

o Other touched files are likely the result of changes to the
  formating or comments.