Chagnes to desgrid calls and formatting
o The call to desgrid_init was moved earlier in the call stack so
  that several des filter checks that depend on the desgrid can
  be added in the correct location.

o The geometry checks for DES were moved into the grid_init routine
  because they are only used to set initial values for the desgrid
  cell widths.

o The max/min radi for discrete particles is set earlier as it is
  needed for setting up the desgrid.

o A first-pass cleanup of the desgrid module was done. More is needed.