Fix for DEM pair data send/recv
Pair data is owned by the particle with the lower index unless the
neighbor particle is a ghost particle. Therefore, when ownership
of a particle is changed, the pair data for that paricle should only
be added if it doesn't already exist. Otherwise, collisions may be
counted multiple times until the next neighbor search is preformed.