Changes to DEM MI layout for cut-cells
The apporach to laying out a DEM MI was changed to keep more MI
cells near to the walls. This was done by intersecting STLs with
the MI mesh. A 'factor of safety' was added by inflating the grid
cells by 10%.

This needs to be added to NS and TB layouts as well.