Incorporated Jeffs changes to fix Issue #34
From Jeff's email to mfix-help on 2-5-2015

Fabio,

I can reproduce the bug in debug mode. There is some out of bound array
access in radial_vel_correction.f. This is the subroutine that circumvents
the singular velocity along the axis of symmetry for cylindrical coordinate
system. Please try with the attached file. I have attempted to fix the
issue, but folks more familiar with this subroutine should take a look at
it. I am copying Akhilesh Bakshi who implemented this correction in MFiX.

...