Extended search for real particles crossing ranks
The previous DES grid IJK value is used to match a real particle
to its ghost counterpart when the ownership of the particle
changes (pack_parcross). However, this can cause problems if the
binning information for the ghost particle was updated. This is
fixed by extending the search to the DES grid cells adjancent to
the previous IJK.

Comment: I'm not sure why the previous IJK value is used in matching
the real particle to the ghost - but - if the current (updated) value
is used, then a similar issue arises. This should only add addional
overhead in the few cases that this issue arises.