Defined DES grid size to fix DMP error
The case would not correctly run in DMP as the DES grid was
much coarser than the fluid grid allowing ghost particles to
exist outside the typical two layers of ghost cells.

A better, long term fix is needed. I think that that PEA should
be modified so that ghost particles can be flaged as outside the
update region for fluid grid quantities. The checks could be done
easily if the bitwise approach outlined in Issue #52 is implemented.