Set FUNIJK_GL to FUNIJK_IO
It appears that at some point the files for reading and converting
data from the binary files were copied into a second file,
file_handle.f. Slight modifications were made to the routines.
These files are no longer called as the files in the model directory,
in_binary_mod, are used.

I'm not sure this modification will fix the broken test cases. If it
does, then the duplicate routines in file_handle, if not the full
file, should be removed.