Remove ghost particles not found on the fluid grid
When the fluid grid an the DES grid are the same size, ghost
particles my be appear outside of the fluid bounds. This is due to
the calculation used for binning the particles. Given that these
particles should not be engaged in collisions, removing them should
not cause any problems.