Changes for cut cells
Do not calculate fluid phase forces for particles that are not in
fluid cells. Particles may legitimately be in a non-fluid cell when
using cut-cells as the fluid domain will not likely match the STL
file.

This change prevents particles in non-fluid cells from calculating
the fluid phase forces to avoid potentially calculating unphysical
values that could cause the simulation to fail. A better fix may
be to assign the particle an IJK index of a neighbor fluid cell.
However, this could cause some issues with interpolation...