Changes to STL binning for wall collisions
The STLs were not getting correctly binned to the DES grid. This
seems to correct the issue. However, additional work is needed
to cleanup the STL preprocessing routines. There are several
arrays that don't appear to be used and should be removed.

Force the STL overlap check for all cells to make sure all
particles are pruned.