Work around for DEM solids inlet
This is a patch that lets a user specify a solids inlet with no
incoming solids (mass flow =  0.0). This would be used when a
simulation starts with no feed but feed is later added. If the
processing isn't done with the NEW run, the solids inlet would
not be available later (unless a RESTART_2 is used).