Modifications to interpolation
The original implementation relied on ghost-particle calculations
to provide data to neighboring processors. This may lead to issues
as the full dataset may be missing information when the interpolation
length is large and the des grid is small. Furthermore, this
approach does not apply to PIC becuase ghost particles are not
used.

Now, data interpolated into ghost cells is added to the 'real' fluid
cell on the neighboring processor.