Adding a new fortran dependency script, bugfixes.
Adding build-aux/fdep.in, a fortran dependecy generator written
in perl. It does the same as the original fortran-depcomp, however
it also:
 - ignores use statements of the form "use, intrinsic"
 - ignores module statements of the form "module procedure"
 - takes into account "include" statements and will recursively
   process the include files.

This means the configure.ac has been updated to find a perl
interpreter (AC_PATH_PROG) with a command line option to
specify it (--with-perl=<LOCATION>).

A new config file rule to generate and change the mode of fdep.

In configure_mfix, added code to copy the generated fdep into
MFiX home (simillar to the makefile code) for out of source
builds.

The makefiles now have an INC variable to specify the include
directories so as to pass them to fdep. They also have a
DEP_IGN variable to specify the "use" statements to ignore in
fdep.

A few of bug fixes.
1) In the out of source build makefile (build-aux/Makefile.usr.am
   changed md5sum to be a variable, that is obtained from
   configure.ac (AC_PATH_PROGS) since it is `md5` on FreeBSD and
   OS X.
2) In both makefiles adding a CLEANS variable to specify a list
   of make goals that we don't want to include/generate the
   dependencies for. The configure.ac "make_ifnotclean" syntax
   has been updated to use "filter" on these goals.
3) Change the recursive chmod call in configure.ac to use "-R"
   as the "--recursive" option is GNU only.
4) Corrected indentation in the Makefile.am libmfix_src to be
   consistent.