Adding array alignment options.
Various preprocessors support the __BIGGEST_ALIGNMENT__ define,
so we will try and use that to future proof the alignment offset
of the start of an array.

This means we have added the following autotools marcos:
1) ax_append_flag.m4 so as to append to the $FCFLAGS variable.
2) ax_fc_check_define.m4 to check if __BIGGEST_ALIGNMENT__ is
   supported.

Changes to the configure.ac include
1) Setting AC_CONFIG_MACRO_DIR() to be ../build-aux so we do not
   need to use m4_include() repeatedly.
2) A call to AX_FC_CHECK_DEFINE() and setting the macro
   __BIGGEST_ALIGNMENT__ to be 64, if it does not exist or is
   smaller than 16.

Changes to model/des/discretelement_mod.f include the addtion
of the compiler directive "ATTRIBUTES ALIGN" so as when arrays
are allocated they are aligned to 64 (in our case) byte boundaries.